requirements.txt
streamlit==1.51.0
langchain==1.0.5
langchain-community==0.4.1
langchain-core==1.0.4
langchain-huggingface==1.0.1
langchain-openai==1.0.2
langchain-text-splitters==1.0.0
langchain-classic==1.0.0
replicate==1.0.7
chromadb==1.3.4
chroma-hnswlib==0.7.3
sentence-transformers==5.1.2
python-dotenv==1.2.1
flask==3.0.0
flask-cors==4.0.0
